+++ packages/urbanjs-tool-tslint/tests/fix-task/gulpfile.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const gulp = require('gulp');
+const tools = require('urbanjs-tools');
+
+tools.tasks.tslint.register(gulp, 'tslint', {
+  files: 'index-invalid.ts'
+});